move `ModuleExtras` out of channel
