+++ crates/ibc/src/test_utils.rs
@@ -2,12 +2,11 @@ use tendermint::{block, consensus, evidence, public_key::Algorithm};
 
 use crate::core::ics04_channel::channel::{Counterparty, Order};
 use crate::core::ics04_channel::error::{ChannelError, PacketError};
-use crate::core::ics04_channel::handler::ModuleExtras;
 use crate::core::ics04_channel::msgs::acknowledgement::Acknowledgement;
 use crate::core::ics04_channel::packet::Packet;
 use crate::core::ics04_channel::Version;
 use crate::core::ics24_host::identifier::{ChannelId, ConnectionId, PortId};
-use crate::core::ics26_routing::context::Module;
+use crate::core::ics26_routing::context::{Module, ModuleExtras};
 use crate::prelude::*;
 use crate::signer::Signer;
